Fix Rubocop offenses in spec/
